Merged Scripts, added Variables
Removed and Merged the Bullet and Enemy "MoveForward" Scripts into the modified "EnemyMoveForward" now known as "EntityMoveForward"; this required adding a Bool Variable referred to as "bool_EntityType" which has "SerializeField" on it, to detect where the object is going, I gave the "InitialVelocity" a "SerializeField" to let it be seen to all things the script was connected to.
Added a check to see what kind of Object was used, which uses the "bool_EntityType" to check, if True the object will go down, and up if False. this also meant I had to modify the Bullet to have the correct values applied to it on top of requiring the Script to be applied to it. Secondly, I have added an additional instance of the SpawnObjectType to it, resulting in the names clarifying the Exact use they will have. these are "SpawnEnemyObject" and "SpawnItemObject".